chore: refactor fee calc in deduct_caller with helpers (#492)
uses `all_mut` so we don't need to fetch values before loading the
account.

It uses `calculate_caller_fee` to check the max balance and deducts the
balance with the effective fee. It is a nice wrapper used for the
mainnet. Additionally, it handles `cfg.is_balance_check_disabled()`
correctly.

Tempo specific fee transfer move to the end.