fix(consensus): allow arbitrarily deep reorgs (#675)
This relaxes the condition for sending a forkchoice update to update the
head hash:

1. Now: set `ForkchoiceState.head_hash` to any `block.hash` as long as
`block.height > last_finalized`.
2. Before: set `ForkchoiceState.head_hash` to `block.hash` if and only
if `block.height > latest_sent_head_height`

So this means the finalized hed acts as the only floor, and above it
arbitary reorgs can happen.

This is done because simplex does not provide any guarantees for
notarized or nullified rounds - only finalized rounds are firmly set.