feat(precompiles): simplify macros (#891)
## Motivation

towards #856 

builds on top of #864 and #872 to simplify the macros leveraging the new
`const: LAYOUT` and the enhanced `trait Storable` with `LayoutCtx`
support. These 2 changes allows us to outsource most of the packing
logic to the types that `impl Storage` and the compiler (thanks to
compile-time const evaluation)

## Solution

refactors the `contract` and `Storable` macros to uses a common
`packing` module that handles the whole slot allocation and constant
generation by leveraging an IR `struct LayoutField` that captures all
important info.

additionally it enhances the `contract`'s built-in constructor method to
perform slot collision checks (optimized away in released builds).