chore(precompiles): drop `SlotId` in favor of runtime slot const (#921)
## Motivation

towards #856 

closes #859 

deprecate the `SlotId` sealed trait in favor of direct usage of
macro-generated constants for each slot.

since slots will only be used by macro-generated fns (not by devs), the
simplification is desirable

## Solution

this simplification also allows us to evaluate `Mapping<K, V>` in the
macro's compile-time constant resolution, which unlocks further
simplifications (on top of #891) when allocationg slots. This is because
mappings become another generic type which is allocated 32 bytes, and
thus cannot be packed.