refactor(tempo-bench): use alloy provider to sign transactions (#1074)
This PR refactors how we sign TIP-20 and DEX transactions in
`tempo-bench`. Previously, we've manually built and signed the
transaction using the signer, now we use Alloy's provider builder
pattern directly on contract instances.

This opens the path to specifying the fee token using
`TempoTransactionRequest`.

Additionally, this removes passing `chain_id` manually and lets Alloy to
figure it out itself.