fix: P256 high s check (#1510)
The secp256k1 function doesn't allow s values higher than n/2, to
prevent 2 tx hashes being valid on the same signature.

We add the same check for our P256 verification function.

The n value for the curve was taken from [RIP
2712](https://eip.tools/rip/7212)
The division with 2 was checked with wolfram, and python scripts.