feat(docs): limit functions that tip-20 fee rule applies to (#1337)
This is a proposed update to the default fee rules. It limits the
functions that, when called on a TIP-20, change the fee token of the
transaction, to only `transfer`, `transferWithMemo`, and
`startReward`—functions that trigger a transfer of tokens from the
caller.

It also limits how this rule is applied to Tempo transactions, so that
it only applies if the `feePayer` of the transaction is the same as
`tx.origin`.

It does not change the rules of `burn` or `burnWithMemo`, despite those
functions triggering transfers from the caller, because those are
administrative functions that are only called by the token's issuer, so
it seems less important. (Open to changing my mind on those.)

Fixes #1112 (since it means that calls to `grantRole` and `mint` no
longer override the default fee token rules).

---------

Co-authored-by: 0xKitsune <77890308+0xKitsune@users.noreply.github.com>